requirements.txt
streamlit
google-generativeai
langchain==0.1.16
langchain-google-genai
langchain-pinecone
pinecone-client
python-dotenv